Delete unnecessary font packages

+++ run.sh
@@ -9,25 +9,24 @@ hostname_laptop="fedora-laptop"
 dnf_apps=(
   ffmpeg gstreamer1-libav util-linux-user fuse-exfat dnf-plugins-core fd-find ripgrep openssl-devel curl wget git
   gnome-tweaks dconf-editor
-  rsms-inter-fonts levien-inconsolata-fonts cascadia-code-fonts mozilla-fira-sans-fonts jetbrains-mono-fonts piper libratbag-ratbagd
-  zsh dash kitty stow btop neofetch ranger spotify-tui
+  rsms-inter-fonts mozilla-fira-sans-fonts
+  libratbag-ratbagd
+  zsh dash kitty stow btop neofetch ranger
   wmctrl wl-clipboard
   code vim-enhanced shellcheck devscripts-checkbashisms
   python3 python3-pip rust cargo java-1.8.0-openjdk java-11-openjdk java-17-openjdk java-latest-openjdk nodejs
-  docker docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose docker-compose-plugin docker-machine
+  docker docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose docker-compose-plugin
 )
 dnf_apps_desktop_only=(
 )
 
 flatpak_apps=(
-  # org.mozilla.firefox
   md.obsidian.Obsidian
   com.transmissionbt.Transmission
   org.gnome.Extensions
   org.gnome.SoundRecorder
   org.gnome.Shotwell # Outdated runtime
   org.gimp.GIMP
-  org.gabmus.hydrapaper # Outdated runtime
   nl.hjdskes.gcolor3
   com.github.tchx84.Flatseal
   com.belmoussaoui.Obfuscate
@@ -48,7 +47,7 @@ flatpak_apps_desktop_only=(
 
 change_hostname() {
   if [ "$1" -eq 1 ]; then
-      hostname="$hostname_desktop"
+    hostname="$hostname_desktop"
   elif [ "$1" -eq 2 ]; then
     hostname="$hostname_laptop"
   fi
@@ -68,7 +67,7 @@ merge_lists() {
 
 update_everything() {
   sudo dnf update -yq
-  sudo dnf upgrade --refresh -yq
+  sudo dnf upgrade -yq --refresh
   flatpak update -y --noninteractive
 }
 
@@ -80,19 +79,19 @@ update_repos_and_apps() {
 install_apps() {
   for app in "${dnf_apps[@]}"; do
     if ! sudo dnf list --installed | grep -q "$app"; then
-      sudo dnf install "$app" -yq
+      sudo dnf install -yq "$app"
       echo "Package $app was installed."
     else
-      echo "Package $app was already installed."
+      echo "Package $app is already installed."
     fi
   done
 
   for app in "${flatpak_apps[@]}"; do
     if ! flatpak list | grep -q "$app"; then
-      flatpak install flathub "$app" -y --noninteractive
+      flatpak install flathub -y --noninteractive "$app"
       echo "Package $app was installed."
     else
-      echo "Package $app was already installed."
+      echo "Package $app is already installed."
     fi
   done
 }
@@ -119,18 +118,15 @@ sh ./components/faster_downloads.sh
 update_everything
 
 # Add VSCode repo
-sudo rpm --import --quiet https://packages.microsoft.com/keys/microsoft.asc
+sudo rpm --quiet --import https://packages.microsoft.com/keys/microsoft.asc
 sudo sh -c 'echo -e "[code]\nname=Visual Studio Code\nbaseurl=https://packages.microsoft.com/yumrepos/vscode\nenabled=1\ngpgcheck=1\ngpgkey=https://packages.microsoft.com/keys/microsoft.asc" > /etc/yum.repos.d/vscode.repo'
 
 # Add RPM Fusion free and nonfree repos
-sudo dnf -yq install https://mirrors.rpmfusion.org/free/fedora/rpmfusion-free-release-"$(rpm -E %fedora)".noarch.rpm https://mirrors.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-"$(rpm -E %fedora)".noarch.rpm
-
-# Add spotify-tui repo (Spotify CLI)
-sudo dnf copr enable atim/spotify-tui -y
+sudo dnf install -yq https://mirrors.rpmfusion.org/free/fedora/rpmfusion-free-release-"$(rpm -E %fedora)".noarch.rpm https://mirrors.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-"$(rpm -E %fedora)".noarch.rpm
 
 # Remove previous Docker versions and add its repo
-sudo dnf -yq remove docker docker-client docker-client-latest docker-common docker-latest docker-latest-logrotate docker-logrotate docker-selinux docker-engine-selinux docker-engine
-sudo dnf -yq config-manager --add-repo https://download.docker.com/linux/fedora/docker-ce.repo
+sudo dnf remove -yq docker docker-client docker-client-latest docker-common docker-latest docker-latest-logrotate docker-logrotate docker-selinux docker-engine-selinux docker-engine
+sudo dnf config-manager -yq --add-repo https://download.docker.com/linux/fedora/docker-ce.repo
 
 # Add flathub repo
 flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
@@ -141,29 +137,32 @@ update_repos_and_apps
 install_apps
 
 # Install neovim as a appimage
-wget -q https://github.com/neovim/neovim/releases/download/stable/nvim.appimage -O "$HOME"/.local/bin/nvim.appimage
+if [ -f "$HOME/.local/bin/nvim.appimage" ]; then
+  echo "Package neovim is already installed."
+else
+  wget -q https://github.com/neovim/neovim/releases/download/stable/nvim.appimage -O "$HOME"/.local/bin/nvim.appimage
+fi
 
 # Update pip before installing everything
-pip install --upgrade pip
+pip install -Uq pip
 
 # Installs TLDR
-pip install -q tldr
+pip install -Uq tldr
 
 # Installs 42 Norminette
-pip install --upgrade pip setuptools
-pip install norminette
-pip install --upgrade norminette
+pip install -Uq setuptools
+pip install -Uq norminette
 
 # Install ruff-lsp to use on neovim
-pip install ruff-lsp
+pip install -Uq ruff-lsp
 
 # Installs black for Python code formatting
-pip install black pynvim
-pip install "black[jupyter]"
+pip install -Uq black pynvim
+pip install -Uq "black[jupyter]"
 
-# Install spotifyd
-cargo install spotifyd --locked
-
+# Install spotifyd and spotify-tui via cargo
+cargo install -q spotifyd --locked
+cargo install -q spotify-tui
 
 # Create Projects folder
 if [ ! -d "$HOME/Projects" ]; then
@@ -173,7 +172,7 @@ fi
 # Configures Gnome to my liking
 sh ./components/config_gnome.sh
 
-# Syncs dotfiles
+# Syncs dotfiles and configs
 sh ./components/sync_dotfiles.sh
 
 # Configures git variables
